Move distributed circuit tests
